Add AgentLoadContext.cs
